Added Keyboard and Mouse interaction, started using more functional states.

+++ src/main/scala/com/colofabrix/scala/tankwar/Configuration.scala
@@ -67,7 +67,7 @@ object Configuration {
     def timeStep: Double = conf.getDouble("simulation.time_step")
 
     /** Target Framerate */
-    def fps: Double = conf.getDouble("simulation.fps")
+    def fps: Int = conf.getInt("simulation.fps")
 
     /** If to display graphics or not */
     def gxfEnabled: Boolean = conf.getBoolean("simulation.gfx_enabled")
